Move logic into SongScrounger class
