+++ run_tests.py
@@ -1,7 +1,6 @@
 import sys
 import unittest
 
-from tests.test_document_parser import TestDocumentParser
 from tests.test_song_scrounger import TestSongScrounger
 from tests.test_spotify_client import TestSpotifyClient
 from tests.test_util import TestUtil
