+++ tests/test_inputs/no_artists_mentioned.txt
@@ -1,0 +1,1 @@
+"American Pie" by someone
