+++ tests/test_inputs/single_artist_mentioned.txt
@@ -1,0 +1,1 @@
+When Don McLean recorded "American Pie"
