+++ tests/test_inputs/test_cross_paragraph_artist_detection.txt
@@ -1,3 +1,3 @@
-Now I'm in some other paragraph over here, mentioning the artist Kiefer.
+The artist that I mention in the next paragraph has a song called "Socially Awkward".
 
-The artist that I mentioned in the previous paragraph has a song called "Island". Even though I don't mention him in this paragraph, I don't mention any other artist that has a song by that name, so it'd be nice for the program to detect him (the guy I mentioned in the last paragraph) as a potential artist.
+Now I'm in some other paragraph over here, mentioning the artist Kiefer.
