--- tests/test_inputs/test_mini.txt
@@ -1,1 +1,0 @@
-When Don McClean recorded "American Pie"
