+++ tests/test_inputs/two_artists_mentioned.txt
@@ -1,0 +1,1 @@
+"American Pie" by Don McLean or... Catch 22 lol
