+++ tests/test_util.py
@@ -7,12 +7,12 @@ from song_scrounger import util
 
 class TestUtil(unittest.TestCase):
     def test_read_file_contents(self):
-        input_file = helper.get_path_to_test_input_file("test_mini.txt")
+        input_file = helper.get_path_to_test_input_file("single_artist_mentioned.txt")
 
         contents = util.read_file_contents(input_file)
 
         self.assertEqual(
-            "When Don McClean recorded \"American Pie\"\n",
+            "When Don McLean recorded \"American Pie\"\n",
             contents,
             "Uexpected file contents!"
         )
